change typed annoted

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -9,7 +9,7 @@ import queue
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: float, max_delay: float) -> float:
+async def wait_n(n: int, max_delay: int) -> float:
     """
     it spawn wait_random n times with the specified
     max_delay
